% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_average_drc.R
\name{akaike_weights}
\alias{akaike_weights}
\title{Akaike weights from a vector of AIC values}
\usage{
akaike_weights(aic)
}
\arguments{
\item{aic}{Numeric vector of AIC values.}
}
\value{
A numeric vector of weights summing to one.
}
\description{
Akaike weights from a vector of AIC values
}
\examples{
akaike_weights(c(100, 102, 110))
}
