% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amphipod_survival}
\alias{amphipod_survival}
\title{Sediment amphipod survival (10 d)}
\format{
A data frame with 35 rows and 4 columns:
\describe{
\item{conc}{Nominal exposure concentration, in the units given below. Zero is the control.}
\item{replicate}{Replicate identifier within concentration.}
\item{total}{Number of individuals exposed in the replicate.}
\item{alive}{Response: surviving amphipods.}
}
}
\source{
Simulated by \code{data-raw/generate_datasets.R}. The parameters used
to generate the data are stored in the \code{truth} attribute of the object,
so that a fitted estimate can be compared against the value that
produced it.
}
\usage{
amphipod_survival
}
\description{
Simulated example data for the \code{amphipod_survival} test type, with a design following USEPA 600/R-94/025. Concentrations are in mg/kg dry sediment.
}
\keyword{datasets}
