% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bayesnec.R
\name{cr_bnec_formula}
\alias{cr_bnec_formula}
\title{Model formula for the bayesnec engine}
\usage{
cr_bnec_formula(test_type, model = NULL)
}
\arguments{
\item{test_type}{Test-type identifier, one of \code{cr_test_types()$id}.}

\item{model}{\code{bayesnec} model name or model-group name. Defaults to the value
in the test-type registry.}
}
\value{
A \code{bayesnecformula} object.
}
\description{
Model formula for the bayesnec engine
}
