% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cr_bundle_outputs}
\alias{cr_bundle_outputs}
\title{Bundle the outputs of one or more analyses into a zip file}
\usage{
cr_bundle_outputs(jobs, zipfile = tempfile(fileext = ".zip"))
}
\arguments{
\item{jobs}{A list of \code{cr_job} objects, or a character vector of file paths.}

\item{zipfile}{Path of the zip file to create.}
}
\value{
The path to the zip file.
}
\description{
Bundle the outputs of one or more analyses into a zip file
}
