% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_types.R
\name{cr_ecx_targets}
\alias{cr_ecx_targets}
\title{ECx levels reported by default for a test type}
\usage{
cr_ecx_targets(id)
}
\arguments{
\item{id}{Test-type identifier, one of \code{cr_test_types()$id}.}
}
\value{
A numeric vector of ECx levels, as percentages.
}
\description{
ECx levels reported by default for a test type
}
