% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cr_job_outputs}
\alias{cr_job_outputs}
\title{Collect the outputs of a finished analysis}
\usage{
cr_job_outputs(job)
}
\arguments{
\item{job}{A \code{cr_job} from \code{\link[=cr_start_job]{cr_start_job()}}.}
}
\value{
A named character vector of the files that exist, as returned by
\code{\link[=cr_output_files]{cr_output_files()}}.
}
\description{
Collect the outputs of a finished analysis
}
