% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cr_stop_job}
\alias{cr_stop_job}
\title{Stop a running analysis}
\usage{
cr_stop_job(job)
}
\arguments{
\item{job}{A \code{cr_job} from \code{\link[=cr_start_job]{cr_start_job()}}.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Stop a running analysis
}
