% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_types.R
\name{cr_test_type}
\alias{cr_test_type}
\title{Look up a single test type}
\usage{
cr_test_type(id)
}
\arguments{
\item{id}{Test-type identifier, one of \code{cr_test_types()$id}.}
}
\value{
A one-row data frame from \code{\link[=cr_test_types]{cr_test_types()}}.
}
\description{
Look up a single test type
}
\examples{
cr_test_type("algal_growth")
}
