% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthworm_reproduction}
\alias{earthworm_reproduction}
\title{Earthworm reproduction (56 d)}
\format{
A data frame with 28 rows and 3 columns:
\describe{
\item{conc}{Nominal exposure concentration, in the units given below. Zero is the control.}
\item{replicate}{Replicate identifier within concentration.}
\item{juveniles}{Response: juveniles per vessel.}
}
}
\source{
Simulated by \code{data-raw/generate_datasets.R}. The parameters used
to generate the data are stored in the \code{truth} attribute of the object,
so that a fitted estimate can be compared against the value that
produced it.
}
\usage{
earthworm_reproduction
}
\description{
Simulated example data for the \code{earthworm_reproduction} test type, with a design following OECD 222. Concentrations are in mg/kg dry soil.
}
\keyword{datasets}
