% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthworm_survival}
\alias{earthworm_survival}
\title{Earthworm acute survival (14 d)}
\format{
A data frame with 28 rows and 4 columns:
\describe{
\item{conc}{Nominal exposure concentration, in the units given below. Zero is the control.}
\item{replicate}{Replicate identifier within concentration.}
\item{total}{Number of individuals exposed in the replicate.}
\item{alive}{Response: surviving adults.}
}
}
\source{
Simulated by \code{data-raw/generate_datasets.R}. The parameters used
to generate the data are stored in the \code{truth} attribute of the object,
so that a fitted estimate can be compared against the value that
produced it.
}
\usage{
earthworm_survival
}
\description{
Simulated example data for the \code{earthworm_survival} test type, with a design following OECD 207. Concentrations are in mg/kg dry soil.
}
\keyword{datasets}
