% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_larval_survival}
\alias{fish_larval_survival}
\title{Fish early life stage survival (7 d)}
\format{
A data frame with 28 rows and 4 columns:
\describe{
\item{conc}{Nominal exposure concentration, in the units given below. Zero is the control.}
\item{replicate}{Replicate identifier within concentration.}
\item{total}{Number of individuals exposed in the replicate.}
\item{alive}{Response: surviving larvae.}
}
}
\source{
Simulated by \code{data-raw/generate_datasets.R}. The parameters used
to generate the data are stored in the \code{truth} attribute of the object,
so that a fitted estimate can be compared against the value that
produced it.
}
\usage{
fish_larval_survival
}
\description{
Simulated example data for the \code{fish_larval_survival} test type, with a design following OECD 210. Concentrations are in mg/L.
}
\keyword{datasets}
