% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{larval_development}
\alias{larval_development}
\title{Larval development (48 h)}
\format{
A data frame with 28 rows and 4 columns:
\describe{
\item{conc}{Nominal exposure concentration, in the units given below. Zero is the control.}
\item{replicate}{Replicate identifier within concentration.}
\item{total}{Number of individuals exposed in the replicate.}
\item{normal}{Response: proportion normal larvae.}
}
}
\source{
Simulated by \code{data-raw/generate_datasets.R}. The parameters used
to generate the data are stored in the \code{truth} attribute of the object,
so that a fitted estimate can be compared against the value that
produced it.
}
\usage{
larval_development
}
\description{
Simulated example data for the \code{larval_development} test type, with a design following ASTM E724. Concentrations are in ug/L.
}
\keyword{datasets}
