% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant_emergence}
\alias{plant_emergence}
\title{Seedling emergence (21 d)}
\format{
A data frame with 28 rows and 4 columns:
\describe{
\item{conc}{Nominal exposure concentration, in the units given below. Zero is the control.}
\item{replicate}{Replicate identifier within concentration.}
\item{sown}{Number of individuals exposed in the replicate.}
\item{emerged}{Response: emerged seedlings.}
}
}
\source{
Simulated by \code{data-raw/generate_datasets.R}. The parameters used
to generate the data are stored in the \code{truth} attribute of the object,
so that a fitted estimate can be compared against the value that
produced it.
}
\usage{
plant_emergence
}
\description{
Simulated example data for the \code{plant_emergence} test type, with a design following OECD 208. Concentrations are in mg/kg dry soil.
}
\keyword{datasets}
