% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cr_data}
\alias{plot_cr_data}
\title{Plot the observed concentration-response data}
\usage{
plot_cr_data(data, test_type, log_x = TRUE)
}
\arguments{
\item{data}{A data frame holding one test.}

\item{test_type}{Test-type identifier, one of \code{cr_test_types()$id}.}

\item{log_x}{Whether to use a logarithmic concentration axis.}
}
\value{
A \code{ggplot} object.
}
\description{
Draws the raw data before any model is fitted. Binomial responses are plotted
as proportions. Zero concentrations are drawn at a position one decade below
the lowest tested concentration and labelled, because a control cannot be
placed on a logarithmic axis but should still be visible.
}
\examples{
data(algal_growth)
plot_cr_data(algal_growth, "algal_growth")
}
