% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_engine}
\alias{require_engine}
\title{Check that an optional fitting engine is installed}
\usage{
require_engine(engine = c("bayesnec", "drc"))
}
\arguments{
\item{engine}{Either \code{"bayesnec"} or \code{"drc"}.}
}
\value{
Invisibly \code{TRUE}, or an error if the package is unavailable.
}
\description{
Check that an optional fitting engine is installed
}
