% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks.R
\name{summarise_design}
\alias{summarise_design}
\title{Summarise the design of a concentration-response dataset}
\usage{
summarise_design(data, test_type)
}
\arguments{
\item{data}{A data frame holding one test.}

\item{test_type}{Test-type identifier, one of \code{cr_test_types()$id}.}
}
\value{
A data frame with one row per concentration giving the number of
replicates and the mean and standard deviation of the response. For
binomial test types the response is summarised as a proportion.
}
\description{
Summarise the design of a concentration-response dataset
}
\examples{
data(algal_growth)
summarise_design(algal_growth, "algal_growth")
}
